requirements.txt
absl-py==2.1.0
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
astunparse==1.6.3
attrs==23.2.0
backtrader==1.9.78.123
blinker==1.8.2
ccxt==4.3.63
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
Cython==3.0.10
dateparser==1.2.0
DateTime==5.5
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.1
gym==0.26.2
gym-notices==0.0.8
h5py==3.11.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
ml-dtypes==0.4.0
mplfinance==0.12.10b0
multidict==6.0.5
namex==0.0.8
norm==1.6.0
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pandas_ta==0.3.14b0
patsy==0.5.6
pillow==10.4.0
playsound==1.3.0
protobuf==4.25.3
pycares==4.4.0
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.18.0
pyparsing==3.1.2
pypiwin32==223
python-binance==1.0.19
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
regex==2024.5.15
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.1
scipy==1.14.0
setuptools==71.0.3
six==1.16.0
statsmodels==0.14.2
ta==0.11.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
win10toast==0.9
wrapt==1.16.0
xgboost==2.1.0
yarl==1.9.4
zope.interface==6.4.post2
